feat(caching): centralize Upstash Redis client configuration (fixes #700)
